+++ components/organisms/UserSettingsPage/user-settings-page.tsx
@@ -10,13 +10,14 @@ import Select from "components/atoms/Select/select";
 import SelectOption from "components/atoms/Select/select-option";
 import LanguagePill from "components/atoms/LanguagePill/LanguagePill";
 
-import { updateUser, updateUserPayload } from "lib/hooks/update-user";
+import { updateUser, UpdateUserPayload } from "lib/hooks/update-user";
 import { ToastTrigger } from "lib/utils/toast-trigger";
 import { authSession } from "lib/hooks/authSession";
 import { validateEmail } from "lib/utils/validate-email";
 import { timezones } from "lib/utils/timezones";
 import { updateEmailPreferences } from "lib/hooks/updateEmailPreference";
 import { useFetchUser } from "lib/hooks/useFetchUser";
+import getInterestOptions from "lib/utils/getInterestOptions";
 
 interface userSettingsPageProps {
   user: User | null;
@@ -42,7 +43,7 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
   });
   const [selectedInterest, setSelectedInterest] = useState<string[]>([]);
   const formRef = useRef<HTMLFormElement>(null);
-  const interestArray = ["javascript", "python", "rust", "ML", "AI", "react"];
+  const interestArray = getInterestOptions();
 
   useEffect(() => {
     async function fetchAuthSession() {
@@ -56,6 +57,8 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
         formRef.current!.twitter_username.value = response.twitter_username;
         formRef.current!.company.value = response.company;
         formRef.current!.location.value = response.location;
+        formRef.current!.github_sponsors_url.value = response.github_sponsors_url;
+        formRef.current!.linkedin_url.value = response.linkedin_url;
       }
     }
     fetchAuthSession();
@@ -116,7 +119,7 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
 
   const handleUpdateProfile = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-    const payload: updateUserPayload = {
+    const payload: UpdateUserPayload = {
       name: formRef.current!.nameInput.value,
       email,
       bio: formRef.current!.bio.value,
@@ -126,9 +129,13 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
       location: formRef.current!.location.value,
       // eslint-disable-next-line camelcase
       display_local_time: displayLocalTime,
-      timezone
+      timezone,
+      // eslint-disable-next-line camelcase
+      github_sponsors_url: formRef.current!.github_sponsors_url.value !== "" ? formRef.current!.github_sponsors_url.value : undefined,
+      // eslint-disable-next-line camelcase
+      linkedin_url: formRef.current!.linkedin_url.value !== "" ? formRef.current!.linkedin_url.value: undefined
     };
-    if(formRef.current?.url.value) {
+    if (formRef.current?.url.value) {
       payload.url = formRef.current!.url.value;
     }
 
@@ -185,8 +192,22 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
               name="url"
             />
             <TextInput
+              classNames="bg-light-slate-4 text-light-slate-11 font-medium"
+              placeholder="https://github.com/sponsors/open-sauced"
+              label="GitHub Sponsors URL"
+              pattern="http[s]?://.*\..{2,}"
+              name="github_sponsors_url"
+            />
+            <TextInput
+              classNames="bg-light-slate-4 text-light-slate-11 font-medium"
+              placeholder="https://www.linkedin.com/in/brianldouglas"
+              label="LinkedIn URL"
+              pattern="http[s]?://.*\..{2,}"
+              name="linkedin_url"
+            />                        
+            <TextInput
               classNames="bg-light-slate-4 text-light-slate-11"
-              placeholder="@saucedopen"
+              placeholder="saucedopen"
               label="Twitter Username"
               name="twitter_username"
             />
@@ -207,7 +228,7 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
                 checked={displayLocalTime}
                 title="profile email"
                 label="Display current local time on profile"
-                onChange={e => setDisplayLocalTime(e.target.checked)}
+                onChange={(e) => setDisplayLocalTime(e.target.checked)}
               />
               <span className="ml-7 text-light-slate-9 text-sm font-normal">
                 Other users will see the time difference from their local time.
@@ -215,7 +236,7 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
             </div>
             <div className="flex flex-col gap-2">
               <label>Time zone*</label>
-              <Select value={timezone} onChange={e => setTimezone(e.target.value)} required>
+              <Select value={timezone} onChange={(e) => setTimezone(e.target.value)} required>
                 <SelectOption value="">Select time zone</SelectOption>
                 {timezones.map((timezone, index) => (
                   <SelectOption key={index} value={timezone.value}>
@@ -224,7 +245,7 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
                 ))}
               </Select>
             </div>
-            <Button disabled={!isValidEmail} type="primary">
+            <Button disabled={!isValidEmail} variant="primary">
               Update profile
             </Button>
           </form>
@@ -245,10 +266,10 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
               ))}
             </div>
             <Button
-              type="default"
+              variant="default"
               disabled={selectedInterest.length === 0}
               onClick={handleUpdateInterest}
-              className="!px-4 !text-light-slate-11 !py-2  !bg-light-slate-4"
+              className="px-4 text-light-slate-11 py-2 bg-light-slate-4"
             >
               Update Interests
             </Button>
@@ -275,8 +296,8 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
             </div>
             <Button
               onClick={handleUpdateEmailPreference}
-              type="default"
-              className="!px-4 w-max !py-2  !bg-light-slate-4 "
+              variant="default"
+              className="px-4 w-max py-2  bg-light-slate-4 "
             >
               Update Preferences
             </Button>